+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - ThumbUpIconNext
   - UsersIconNext
   - WindowEditIconNext
+- Fix: `ButtonWithTooltip`: Prevent unwanted animation on load.
 - Fix: `Alert`: Fix responsive styles.
 
 ## [5.3.0] - 2021-10-19